Modified the file

+++ control-flow/match_case_calculator.py
@@ -1,29 +1,24 @@
 #Using match case statements to handle arithmetic operations
-num1 = int(input("Enter the first number: "))
-num2 = int(input("Enter the second number: "))
+num1 = float(input("Enter the first number: "))
+num2 = float(input("Enter the second number: "))
 operator = input("Choose the operation (+, -, *, /): ")
 
 match operator:
-    case "+":
+    case '+':
         result = num1 + num2
         print(f"The result is {result}")
-
-    case "-":
+    case '-':
         result = num1 - num2
         print(f"The result is {result}")    
-
-    case "*":
+    case '*':
         result = num1 * num2
         print(f"The result is {result}")    
-
-    case "/":
+    case '/':
         if num2 == 0: #Division by zero check
-                
-            print("Cannot divide by zero")
+              print("Cannot divide by zero")
         else:
             result = num1 / num2
             print(f"The result is {result}")
-    
     case _:
         print("Invalid operation")
